Stop fetching all available chromedriver versions except the latest
Since `google-chrome` is an evergreen browser, there is no assurance of
support for any version other than the latest. This means that there
shouldn't be any need (theoretically) to support old `chromedriver`
versions as well. Given this, we could do away with the dependency on
`nokogiri` entirely by just considering / fetching the latest version
number.

Installing custom / older versions is still supported for backwards
compatibility but the manifests no longer check if the custom version
is valid and will probably fail with a cryptic error message related to
a failure to download the `chromedriver` archive